Refactores queue_subscriber module
