+++ src/queues/mod.rs
@@ -3,7 +3,6 @@ mod queue_data;
 
 mod queue_metrics;
 mod queues_list;
-pub mod subscribers;
 
 pub use queue::TopicQueue;
 pub use queue_data::{NextMessage, QueueData};
